Allow optimization of binary matches in escapeu_1/8 and escapeu_2/8
This quiets the 'NOT OPTIMIZED' messages from the compilation
with binary optimization messages:

     ERL_COMPILER_OPTIONS=bin_opt_info rebar3 compile

The compiler doesn't have enough information in the body of
`escapeu_1/8` or `escapeu_2/8` to know that `Rest` will always
be a bitstring. We can tell it so in the parameter list with a
pattern match on any bitstring.

Although it doesn't appear to have a noticeable effect on speed
according to the benchmark.